feat: update introjs to new daisyui
