--- assets/js/intro-init.js
@@ -1,2 +1,0 @@
-import introJs from 'intro.js';
-globalThis.introJs = introJs;
