+++ website/templates/calendar.j2
@@ -15,5 +15,5 @@
 {% endblock %}
 {% block scripts %}
 {{ super() }}
-<script src="{{ asset('assets/js/calendar.js') }}"></script>
+<script type="module" src="{{ asset('assets/js/calendar.js') }}"></script>
 {% endblock %}
